Add tool to wait until replication is finished
Since it reuses the same information retrieved by sep-replication-status, refactored the code so that it is more modular.
Note that this tool is not perfect, it is based on heuristics to determine if the replication is done: no hlog files queued (beyond the current one) and no activity during the last two minutes.