Base ForkedReplicationSource on hbase 0.94.2
Note that while we depend on CDH 4.2.0, which contains HBase 0.94.2, we use the ReplicationSource from a newer version.
Main reason to do so is that with 0.94.2, I saw recovered queues that were not cleaned up in case the queue consisted of empty hlog files (which you get when restarting an hbase cluster that didn't receive any activity). These recovered queues then stay around forever, with correponding background threads running. I haven't studied in detail what the changes are, only noticed that this version of the code didn't apppear to have the problem.