demo: create splitted table + appropriate row keys
To make the demo more interesting on a real cluster, create a pre-splitted table so that the whole cluster will be replicating from the beginning.
Removed dashes from the default uuid.tostring representation so that they don't appear in the part of the string that is used for split selection.