Separate Indexer from SepListener
Extract the indexing functionality of the Indexer class, and
add a new class that is a SepListener and passes data down to
an underlying Indexer.

This commit also introduces the RowData interface, which has
implementations that wrap SepEvents or Results. The Indexer
functionality runs based on RowData instances, allowing it to be
used in both NRT and batch contexts.