Fix NPE if ResultToSolrMapper returns null
The ResultToSolrMapper interface specifies that the map(Result)
method can return null if there is nothing to be indexed, but the
Indexer implementations were not handling the null properly and
were throwing NullPointerException. This fix handles the null
value correctly as specified by the ResultToSolrMapper (i.e. a
null SolrInputDocument is simply discarded).